pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autoservice"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "beautifulsoup4>=4.14.3",
    "claude-agent-sdk>=0.1.21",
    "fastapi>=0.109.0",
    "httpx>=0.28.1",
    "lark-oapi>=1.3.0",
    "mcp>=1.0.0",
    "openpyxl>=3.1.5",
    "pypdf>=6.7.4",
    "python-docx>=1.2.0",
    "uvicorn[standard]>=0.27.0",
    "pyyaml>=6.0",
    "anyio>=4.0",
    "websockets>=13.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
